[refactor] s3.tx에서 getKey 함수 분리, 함수명 변경
